Linux Scripts/SCP/requirements.txt
scp
paramiko
pyyaml
cryptography==2.4.2
